Error handling is now implemented correctly, such that the interpreter can continue running after an error without losing state.  In order for this to work though, error checking had to be moved out of the stils and stilo_* classes, directly into the code using the various interfaces, so that the caller could clean up and return without doing any permanent work.  This conversion is complete for all use of the stils class, but has yet to be done for the stilo_* classes.
From a performance perspective, this should speed things up, because there
are a number of places inside tight loops that the interpreter knows certain
errors won't happen, so it doesn't check any more.  From a size perspective,
this increases the size of the library significantly, since error checking
that used to be done in a handful of places is now done in hundreds of
places.  Oh well, this is the correct way to do it, which is most important.