[Ignoring previous revision.]
Modify log structure such that all records have a 9 byte header.  This makes
it possible to have insert/yank/remove/delete records of arbitrary size,
without the need to break the records up.  Also, the position record now
stores the position in the record header, thus removing the need for a
payload in position records.

The eventual goal of these changes is to vastly reduce the overhead for
large insert/yank/remove/delete operations.  Part of making that possible
is changing the data byte ordering in insert/yank records so that the data
can be used directly when making buf API calls.  However, we have bufv_rcopy(),
so it may make more sense to revert the log data byte ordering, and fix
things up after the fact when undoing/redoing.  The current result though is
that undo reverses data right now, and that needs to be fixed one way or
another.