Fix various bugs introduced by reverse insertion.
Invert the from/to loops in bufv_rcopy() and bufv_p_rcopy() so that the
destination is filled from the beginning, rather from the end, so that any
leftover space is at the end.

Significantly modify mkr_range_get() to avoid ever creating bufv elements
that have 0 length, and add assertion code to check this.